requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
icalendar==6.0.0
python-dateutil==2.9.0
pyyaml==6.0.2
beautifulsoup4==4.12.3
lxml==5.3.0
pytz==2024.2
